refactor(parse): split manifest parsers by format
